+++ source/extensions/isaacsim.sensors.camera/docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.2.4] - 2024-11-26
+### Fixed
+- camera view sensor test warp.types.int32 -> warp.types.uint32
+- decreased image comparison threshold with 0.95->0.94
+
 ## [0.2.3] - 2024-11-26
 ### Fixed
 - Camera sensor tests fix for colorize param